# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: huMANity API - Explore history and culture of nations
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [connections]
      summary: Search people by name, username, or email
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching people with the viewer's connection status to each
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSearchResult"
        "401":
          description: Unauthorized
  /connections/recommendations:
    get:
      operationId: getConnectionRecommendations
      tags: [connections]
      summary: Recommend people to connect with based on shared interests
      description: Ranks non-connected members by overlap of interests, languages, country, and favorites with the signed-in user.
      responses:
        "200":
          description: Recommended people, highest match first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecommendedUser"
        "401":
          description: Unauthorized
  /connections:
    get:
      operationId: listConnections
      tags: [connections]
      summary: List the signed-in user's accepted Human Connections
      responses:
        "200":
          description: Accepted connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConnectionUser"
        "401":
          description: Unauthorized
  /connections/requests:
    get:
      operationId: listConnectionRequests
      tags: [connections]
      summary: List pending incoming and outgoing connection requests
      responses:
        "200":
          description: Pending requests grouped by direction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionRequestsList"
        "401":
          description: Unauthorized
    post:
      operationId: sendConnectionRequest
      tags: [connections]
      summary: Send a connection request to another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionRequestInput"
      responses:
        "201":
          description: Resulting connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionStatusResult"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: Target user not found
  /connections/requests/{id}/accept:
    post:
      operationId: acceptConnectionRequest
      tags: [connections]
      summary: Accept an incoming connection request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resulting connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionStatusResult"
        "401":
          description: Unauthorized
        "404":
          description: Request not found
  /connections/requests/{id}:
    delete:
      operationId: deleteConnectionRequest
      tags: [connections]
      summary: Decline an incoming request or cancel an outgoing one
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Request removed
        "401":
          description: Unauthorized
        "404":
          description: Request not found
  /connections/status/{userId}:
    get:
      operationId: getConnectionStatus
      tags: [connections]
      summary: Get the viewer's connection status with a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionStatusResult"
        "401":
          description: Unauthorized
  /connections/{userId}:
    delete:
      operationId: removeConnection
      tags: [connections]
      summary: Remove an existing connection with a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Connection removed
        "401":
          description: Unauthorized
        "404":
          description: Connection not found
...
components:
  schemas:
...
    ConnectionUser:
      type: object
      required: [userId, displayName, humanityScore, pledged]
      properties:
        userId:
          type: string
        username:
          type: string
          nullable: true
        displayName:
          type: string
        photoUrl:
          type: string
          nullable: true
        countryName:
          type: string
          nullable: true
        countryFlagUrl:
          type: string
          nullable: true
        humanityScore:
          type: integer
        pledged:
          type: boolean
    UserSearchResult:
      type: object
      required: [userId, displayName, humanityScore, pledged, connectionStatus]
      properties:
        userId:
          type: string
        username:
          type: string
          nullable: true
        displayName:
          type: string
        photoUrl:
          type: string
          nullable: true
        countryName:
          type: string
          nullable: true
        countryFlagUrl:
          type: string
          nullable: true
        humanityScore:
          type: integer
        pledged:
          type: boolean
        connectionStatus:
          type: string
          enum: [none, pending_outgoing, pending_incoming, connected]
        connectionId:
          type: integer
          nullable: true
    RecommendedUser:
      type: object
      required: [userId, displayName, humanityScore, pledged, matchScore, sharedInterests, sharedLanguages, sharedCountry, reasons]
      properties:
        userId:
          type: string
        username:
          type: string
          nullable: true
        displayName:
          type: string
        photoUrl:
          type: string
          nullable: true
        countryName:
          type: string
          nullable: true
        countryFlagUrl:
          type: string
          nullable: true
        humanityScore:
          type: integer
        pledged:
          type: boolean
        matchScore:
          type: integer
        sharedInterests:
          type: array
          items:
            type: string
        sharedLanguages:
          type: array
          items:
            type: string
        sharedCountry:
          type: boolean
        reasons:
          type: array
          items:
            type: string
...
    ConnectionRequestItem:
      type: object
      required: [id, createdAt, user]
      properties:
        id:
          type: integer
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/ConnectionUser"
    ConnectionRequestsList:
      type: object
      required: [incoming, outgoing]
      properties:
        incoming:
          type: array
          items:
            $ref: "#/components/schemas/ConnectionRequestItem"
        outgoing:
          type: array
          items:
            $ref: "#/components/schemas/ConnectionRequestItem"
    ConnectionStatusResult:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [none, pending_outgoing, pending_incoming, connected]
        connectionId:
          type: integer
          nullable: true
    ConnectionRequestInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
...