# lib/api-zod/src/generated/types/connectionStatusResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * huMANity API - Explore history and culture of nations
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectionStatusResultStatus } from './connectionStatusResultStatus';

export interface ConnectionStatusResult {
  status: ConnectionStatusResultStatus;
  connectionId?: number | null;
}